# docs/06-evidence/users/auth_users_openapi.json
{
  "components": {
    "schemas": {
...
      "WorkflowAvailability": {
        "properties": {
          "status": {
            "enum": [
              "available",
              "degraded"
            ],
            "title": "Status",
            "type": "string"
          },
          "unavailableTargetCount": {
            "title": "Unavailabletargetcount",
            "type": "integer"
          }
        },
        "required": [
          "status",
          "unavailableTargetCount"
        ],
        "title": "WorkflowAvailability",
        "type": "object"
      },
      "WorkflowCreate": {
        "additionalProperties": false,
        "properties": {
          "confirmed": {
            "const": true,
            "title": "Confirmed",
            "type": "boolean"
          },
          "description": {
            "anyOf": [
              {
                "maxLength": 600,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "sourceRef": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sourceref"
          },
          "sourceType": {
            "enum": [
              "agent",
              "builtin",
              "manual"
            ],
            "title": "Sourcetype",
            "type": "string"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/WorkflowStepInput"
            },
            "maxItems": 30,
            "minItems": 1,
            "title": "Steps",
            "type": "array"
          },
          "title": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title",
          "sourceType",
          "confirmed",
          "steps"
        ],
        "title": "WorkflowCreate",
        "type": "object"
      },
      "WorkflowCreateMeta": {
        "properties": {
          "contractVersion": {
            "title": "Contractversion",
            "type": "integer"
          },
          "idempotencyReplayed": {
            "title": "Idempotencyreplayed",
            "type": "boolean"
          }
        },
        "required": [
          "idempotencyReplayed",
          "contractVersion"
        ],
        "title": "WorkflowCreateMeta",
        "type": "object"
      },
      "WorkflowCreateResponse": {
        "properties": {
          "meta": {
            "$ref": "#/components/schemas/WorkflowCreateMeta"
          },
          "workflow": {
            "$ref": "#/components/schemas/WorkflowResponseItem"
          }
        },
        "required": [
          "workflow",
          "meta"
        ],
        "title": "WorkflowCreateResponse",
        "type": "object"
      },
      "WorkflowEnvelope": {
        "properties": {
          "workflow": {
            "$ref": "#/components/schemas/WorkflowResponseItem"
          }
        },
        "required": [
          "workflow"
        ],
        "title": "WorkflowEnvelope",
        "type": "object"
      },
      "WorkflowListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/WorkflowResponseItem"
            },
            "title": "Items",
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/WorkflowMeta"
          }
        },
        "required": [
          "items",
          "meta"
        ],
        "title": "WorkflowListResponse",
        "type": "object"
      },
      "WorkflowMeta": {
        "properties": {
          "contractVersion": {
            "title": "Contractversion",
            "type": "integer"
          },
          "hasNext": {
            "title": "Hasnext",
            "type": "boolean"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "pageSize": {
            "title": "Pagesize",
            "type": "integer"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "totalCount": {
            "title": "Totalcount",
            "type": "integer"
          }
        },
        "required": [
          "page",
          "pageSize",
          "totalCount",
          "hasNext",
          "source",
          "contractVersion"
        ],
        "title": "WorkflowMeta",
        "type": "object"
      },
      "WorkflowResponseItem": {
        "properties": {
          "archivedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archivedat"
          },
          "availability": {
            "$ref": "#/components/schemas/WorkflowAvailability"
          },
          "createdAt": {
            "title": "Createdat",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "sourceRef": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sourceref"
          },
          "sourceType": {
            "enum": [
              "agent",
              "builtin",
              "manual"
            ],
            "title": "Sourcetype",
            "type": "string"
          },
          "status": {
            "enum": [
              "active",
              "archived"
            ],
            "title": "Status",
            "type": "string"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/WorkflowStepResponse"
            },
            "title": "Steps",
            "type": "array"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updatedAt": {
            "title": "Updatedat",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          },
          "workflowUid": {
            "title": "Workflowuid",
            "type": "string"
          }
        },
        "required": [
          "workflowUid",
          "title",
          "description",
          "sourceType",
          "sourceRef",
          "status",
          "version",
          "archivedAt",
          "createdAt",
          "updatedAt",
          "steps",
          "availability"
        ],
        "title": "WorkflowResponseItem",
        "type": "object"
      },
      "WorkflowStatusUpdate": {
        "additionalProperties": false,
        "properties": {
          "expectedVersion": {
            "minimum": 1.0,
            "title": "Expectedversion",
            "type": "integer"
          }
        },
        "required": [
          "expectedVersion"
        ],
        "title": "WorkflowStatusUpdate",
        "type": "object"
      },
      "WorkflowStepInput": {
        "additionalProperties": false,
        "properties": {
          "name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "objective": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Objective",
            "type": "string"
          },
          "order": {
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Order",
            "type": "integer"
          },
          "toolHrefSnapshot": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Toolhrefsnapshot"
          },
          "toolNameSnapshot": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Toolnamesnapshot"
          },
          "toolSlug": {
            "anyOf": [
              {
                "maxLength": 100,
                "minLength": 1,
                "pattern": "^[A-Za-z0-9_-]+$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Toolslug"
          }
        },
        "required": [
          "order",
          "name",
          "objective"
        ],
        "title": "WorkflowStepInput",
        "type": "object"
      },
      "WorkflowStepResponse": {
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "objective": {
            "title": "Objective",
            "type": "string"
          },
          "stepOrder": {
            "title": "Steporder",
            "type": "integer"
          },
          "stepUid": {
            "title": "Stepuid",
            "type": "string"
          },
          "target": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WorkflowTarget"
              },
              {
                "type": "null"
              }
            ]
          },
          "toolHrefSnapshot": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Toolhrefsnapshot"
          },
          "toolNameSnapshot": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Toolnamesnapshot"
          },
          "toolSlug": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Toolslug"
          }
        },
        "required": [
          "stepUid",
          "stepOrder",
          "name",
          "objective",
          "toolSlug",
          "toolNameSnapshot",
          "toolHrefSnapshot",
          "target"
        ],
        "title": "WorkflowStepResponse",
        "type": "object"
      },
      "WorkflowTarget": {
        "properties": {
          "href": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Href"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "status": {
            "enum": [
              "available",
              "unavailable"
            ],
            "title": "Status",
            "type": "string"
          },
          "type": {
            "const": "tool",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "key",
          "status",
          "name",
          "href"
        ],
        "title": "WorkflowTarget",
        "type": "object"
      },
      "WorkflowUpdate": {
        "additionalProperties": false,
        "properties": {
          "description": {
            "anyOf": [
              {
                "maxLength": 600,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "expectedVersion": {
            "minimum": 1.0,
            "title": "Expectedversion",
            "type": "integer"
          },
          "title": {
            "anyOf": [
              {
                "maxLength": 120,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "required": [
          "expectedVersion"
        ],
        "title": "WorkflowUpdate",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Auth Users Contract Baseline",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/users/me/assets/workflows": {
      "get": {
        "operationId": "list_workflows_api_v1_users_me_assets_workflows_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "pageSize",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Pagesize",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "default": "active",
              "enum": [
                "active",
                "archived",
                "all"
              ],
              "title": "Status",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Workflows",
        "tags": [
          "user-assets"
        ]
      },
      "post": {
        "operationId": "create_workflow_api_v1_users_me_assets_workflows_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 8,
              "title": "Idempotency-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowCreateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Workflow",
        "tags": [
          "user-assets"
        ]
      }
    },
    "/api/v1/users/me/assets/workflows/{workflow_uid}": {
      "get": {
        "operationId": "get_workflow_api_v1_users_me_assets_workflows__workflow_uid__get",
        "parameters": [
          {
            "in": "path",
            "name": "workflow_uid",
            "required": true,
            "schema": {
              "title": "Workflow Uid",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Workflow",
        "tags": [
          "user-assets"
        ]
      },
      "patch": {
        "operationId": "update_workflow_api_v1_users_me_assets_workflows__workflow_uid__patch",
        "parameters": [
          {
            "in": "path",
            "name": "workflow_uid",
            "required": true,
            "schema": {
              "title": "Workflow Uid",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Workflow",
        "tags": [
          "user-assets"
        ]
      }
    },
    "/api/v1/users/me/assets/workflows/{workflow_uid}/archive": {
      "post": {
        "operationId": "archive_workflow_api_v1_users_me_assets_workflows__workflow_uid__archive_post",
        "parameters": [
          {
            "in": "path",
            "name": "workflow_uid",
            "required": true,
            "schema": {
              "title": "Workflow Uid",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowStatusUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Archive Workflow",
        "tags": [
          "user-assets"
        ]
      }
    },
    "/api/v1/users/me/assets/workflows/{workflow_uid}/restore": {
      "post": {
        "operationId": "restore_workflow_api_v1_users_me_assets_workflows__workflow_uid__restore_post",
        "parameters": [
          {
            "in": "path",
            "name": "workflow_uid",
            "required": true,
            "schema": {
              "title": "Workflow Uid",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowStatusUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Restore Workflow",
        "tags": [
          "user-assets"
        ]
      }
    },
...
  }
}